pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-mcp"
version = "0.1"
description = "让 MCP 工具像本地函数一样简单的桥接库"
authors = [
    { name = "" }
]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "langchain-core",
    "pydantic",
    "mcp",
    "python-dotenv"
]

[tool.setuptools.package-dir]
"" = "src"

# 可以在后续添加实际的URL
# [project.urls]
# "Homepage" = "https://example.com"
# "Bug Tracker" = "https://example.com/issues"